pyproject.toml
[tool.poetry]
name = "finance-bot"
version = "0.1.0"
description = ""
authors = ["ivan-karaberov"]
readme = "README.md"
packages = [{include = "finance_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.1.1"
flake8 = "^6.1.0"
jinja2 = "^3.1.2"
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
